Split up the device deployment
I think this is better as two goals. The build goal is usable on its own for things like automatic
IPA upload. The device deployment is more appropriate for local debugging but requires a build in
order to work.

Also consolidate the code to build the app, since it's the same for all 3 mojos.