Allow projects to be nested.
This shouldn't affect projects with the standard layout, but will allow my project to work. It
calculates the location of the app relative to the location of the solution. My setup has a csproj
per deployment, within a directory. So the app is in e.g. staging/bin, not ./bin.